feat: refactor ZeWeatherService to use DI
